refactor(web-terminal): one module owns the docs/feedback defaults
The three "who owns this deployment" defaults each lived in two places, and
the second copy was hand-typed: routes/panels.py re-spelled the docs URL, the
maintainer address and — worse — the github_repo sugar expansion as literals
in its getattr fallbacks. routes/feedback.py did the same for the store
ceiling, a fourth site.

The comment blamed a routes -> app import cycle. The cycle is real for a
top-level `from ...app import` (the app pulls the router in at import time),
but it was never a reason to duplicate a value: a sibling leaf beside
feedback_composer and feedback_store is reachable from both sides.

So feedback_destination.py now owns the constants and the whole coercion
family that only ever served them — coerce_config_str, the tracker coercions,
resolve_feedback_trackers and coerce_store_ceiling. app.py and both route
modules import it; every literal is deleted.

The duplicated copies sat on the FALLBACK path, which only runs when app
state is missing, so drift between the two spellings was invisible in every
ordinary deployment and would have surfaced once, in the degraded case
nobody is watching.

No behavior change.
